# local-api-docs.yaml
openapi: 3.0.1
info:
  title: gateways-api
  description: a private microservice that serves as gateways and devices
  version: v1.0.0
servers:
  - url: http://localhost:8080/gateways-api
    description: Generated server url
security:
  - basicAuth: []
paths:
  /api/rest/{gateway_uuid}/device/{device_uuid}:
    get:
      tags:
        - getDevice
      summary: get a gateway's device
      operationId: getDevice
      parameters:
        - name: gateway_uuid
          in: path
          description: Uuid of the gateway that contains wanted device. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3c221f5-3c2d-11ec-a662-0242ac160003
        - name: device_uuid
          in: path
          description: Uuid of the device to be read. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3e28a87-3c2d-11ec-a662-0242ac160003
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad device_uuid, not in uuid format.:
                  description: "An example Response to a bad device_uuid, not in uuid\
                    \ format."
                  value:
                    error: "Validation failed, getDevice.gateway_uuid: Requested uuid\
                      \ q3c23161-3c2d-11ec-a662-0242ac160003 is not valid."
        "404":
          description: Requested resource is not found
          content:
            application/json:
              examples:
                An example Response to device_uuid not found.:
                  description: An example Response to device_uuid not found.
                  value:
                    error: Requested Device with uuid ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                      is not Found
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:45:26.377+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f/device/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
        "200":
          description: successfully retrieved a single Device entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModel'
              examples:
                An example Response to getDevice endpoint.:
                  description: An example Response to getDevice endpoint.
                  value:
                    devicesUuid: ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                    devicesName: Device-3ebae3e5-bd64-4c8e-8370-48f96717e9e8
                    vendor: texas tech
                    deviceCreationDate: 2021-11-16T03:08:32.258
                    status: true
    put:
      tags:
        - updateDevice
      summary: update a Device details
      operationId: updateDevice
      parameters:
        - name: gateway_uuid
          in: path
          description: Uuid of the gateway that contains wanted device. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3c221f5-3c2d-11ec-a662-0242ac160003
        - name: device_uuid
          in: path
          description: Uuid of the device to be updated. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3e28a87-3c2d-11ec-a662-0242ac160003
      requestBody:
        description: Device to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceModel'
            example:
              devicesUuid: a3e3befe-3c2d-11ec-a662-0242ac160003
              devicesName: device-a3e3bf16-3c2d-11ec-a662-0242ac160003
              vendor: texas tech
              deviceCreationDate: 2021-11-02T22:38:46
              status: true
        required: true
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad device_uuid, not in uuid format.:
                  description: "An example Response to a bad device_uuid, not in uuid\
                    \ format."
                  value:
                    error: "Validation failed, updateDevice.gateway_uuid: Requested\
                      \ uuid q3c23161-3c2d-11ec-a662-0242ac160003 is not valid."
        "404":
          description: Requested resource is not found
          content:
            application/json:
              examples:
                An example Response to device_uuid not found.:
                  description: An example Response to device_uuid not found.
                  value:
                    error: Requested Device with uuid ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                      is not Found
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "200":
          description: Gateway's Device updated successfully
          content:
            application/json:
              examples:
                An example Response to updateDevice endpoint.:
                  description: An example Response to updateDevice endpoint.
                  value:
                    message: Gateway's Device updated successfully
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:53:05.949+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f/device/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
    delete:
      tags:
        - deleteDevice
      summary: delete a Device details
      operationId: deleteDevice
      parameters:
        - name: gateway_uuid
          in: path
          description: Uuid of the gateway that contains wanted device. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3c221f5-3c2d-11ec-a662-0242ac160003
        - name: device_uuid
          in: path
          description: Uuid of the device to be deleted. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3e28a87-3c2d-11ec-a662-0242ac160003
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad device_uuid, not in uuid format.:
                  description: "An example Response to a bad device_uuid, not in uuid\
                    \ format."
                  value:
                    error: "Validation failed, deleteDevice.gateway_uuid: Requested\
                      \ uuid q3c23161-3c2d-11ec-a662-0242ac160003 is not valid."
        "404":
          description: Requested resource is not found
          content:
            application/json:
              examples:
                An example Response to device_uuid not found.:
                  description: An example Response to device_uuid not found.
                  value:
                    error: Requested Device with uuid ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                      is not Found
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:53:05.949+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f/device/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
        "200":
          description: Gateway's Device deleted successfully
          content:
            application/json:
              examples:
                An example Response to deleteDevice endpoint.:
                  description: An example Response to deleteDevice endpoint.
                  value:
                    message: Gateway's Device deleted successfully
...
components:
  schemas:
    DeviceModel:
      type: object
      properties:
        devicesUuid:
          type: string
        devicesName:
          type: string
        vendor:
          type: string
        deviceCreationDate:
          type: string
          format: date-time
        status:
          type: boolean
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic